Bug#11827404:60114: INSERT DELAYED CAN CALL HA_RELEASE_AUTO_INCREMENT                     WITHOUT HAVING A LOCK.
Analysis
-------

Mysqld crashes at times on a debug build, during the execution of
the INSERT DELAYED statement under error scenarios. When a thread
executes a DELAYED statement for a table, a handler thread is
created to process all DELAYED statements for the table. The handler
thread locks the table and handles the queued up inserts on the table.
The handler thread updates the auto increment value row prior to
writing the record. After the handler thread processes the stacked
inserts, the auto increment value is reset and the table is unlocked.
In case of simple INSERT, the thread that received INSERT request
resets the auto increment value after writing the record. But not
in the case of DELAYED INSERT since the threads share the table
object.

Under error scenarios, the thread that received the INSERT DELAYED
request makes a redundant call to reset the auto increment value.

The crash is observed in an error scenario and under a timing window
where:
The handler thread resets the auto increment value and unlocks
the table, then the thread which received the INSERT DELAYED request
tries to call the redundant release of auto increment value without
having the lock on the table.

Fix:
---
Since the handler thread ensures the cleanup, removed the redundant
call to reset the auto increment value in the main thread. Also in
case of simple INSERT, it was observed that the call to reset auto
increment value was redundant since the release was performed much
ahead during the execution.

NOTE:No regression test case has been added because the sync points
currently does not work for the handler thread. The patch has
been verified by running the existing test case main.delayed which
fails sometimes without this patch.