WL#6082 Improve the Disk-Sweep Multi-Range Read cost model
Patch 5: IO cost estimate is too high compared to "standard MRR"

The DS-MRR cost model has two different ways for calculating the
IO-cost of doing a "disk sweep scan":

a. Disk-sweep model: Used when the data is not stored as a
   clustered index: The model estimates the number of blocks that will
   be accessed and models the time to read these using sequential
   scan/read of a data file. This model is used for MyISAM.

b. If the data is stored as a clustered index: The cost model
   estimates this as having to do one random block access 
   per record. This model does not take into account that
   we actually will be accessing the table as a scan. This
   model is used for InnoDB.

The focus is on improving the model for InnoDB since this does not
take into account any benefit from sequential access of the table. Due
to this, the IO-cost estimate will in almost all cases be larger for
DS-MRR than "standard MRR" (and thus, the cost based DS-MRR will not
be used for InnoDB).

The fix for this problem is to switch to use the "disk streaming IO
cost model" (as described as a above) also when the data is stored in
a clustered index.