Followup to WL#6189 Turn InnoDB persistent statistics ON by default
Adjust mtr tests, part 27.

Persistent stats use a different sampling algorithm so it is possible
that the stats numbers differ from transient stats.

Also, persistent stats are updated less frequently or with a delay, so
it is possible that persistent stats are not up to date as transient were
even if both algorithms would return the same results.

The order of the returned rows by a SELECT query depends on the stats and
thus, if possible, we fix tests by prepending "-- sorted_result" to SELECTs
that happen to return rows in different order.

If possible, each failing test was fixed by manually running ANALYZE TABLE.
This is doable if both transient and persistent sampling algorithms end up
with the same numbers for the given table and its data.

If persistent stats result in different stats, then test failures were fixed
by forcing the usage of transient stats for the table by using
CREATE TABLE ... STATS_PERSISTENT=0.

Intentionally do not fix the tests by using persistent stats and adjustin
the output of EXPLAIN in .result files because a different execution plan
may cause a different code path to be executed, than the one originally
intended in the test.