Bug#11765737: USE_COUNT: WRONG COUNT FOR KEY AT 0X7FFFD423EDE0,               0 SHOULD BE 1
SEL_ARG::use_count is used to keep track of the number of 
times a SEL_ARG is referred. SEL_ARGs are e.g. referred
from SEL_ARGs with an earlier key part number. This is done so 
that multiple range predicates with equal subparts can use the
same SEL_ARG instead of duplicating the effort. If an index is
created with two keyparts (kp1, kp2) and the query contains the
predicate "(kp1=1 or kp1=2) and (kp2>3)", two SEL_ARGs needs
to be created for the kp1 part of the predicate. Both these 
SEL_ARGs will have use_count=1. Furthermore, both will point 
to the SEL_ARG for kp2 which will have use_count=2.

The bug was in SEL_ARG::increment_use_count() which is called
on a SEL_ARG when more SEL_ARGs start referring to it. When
increment_use_count() was called on a SEL_ARG, the SEL_ARGs
it referred to was incremented by the input increment multiplied
by the SEL_ARGs previous use_count. For a predicate
"(kp1=1 or kp1=2 or kp1=3) and (kp2>3) and kp3>4" that means 
that when these SEL_ARG trees get merged (use count numbers in
brackets):
"(kp1=1[1] or kp1=2[1]) and kp2>3[2] and kp3>4[2]" 
"kp1=3[1]               and kp2>3[1] and kp3>4[1]",
increment_use_count() would add 1 to use_count of kp2 and 2 to 
use_count of kp3, resulting in this SEL_ARG tree:
"(kp1=1[1] or kp1=2[1] or kp1=2[1]) and kp2>3[3] and kp3>4[4]"

The use_count of kp3 should have been 3. The fix is to remove
the multiplication of use_count in 
SEL_ARG::increment_use_count().