ndb - fifo idle connection list
This ensures that the list of connections to the TC block is kept in a FIFO queue rather than a
LIFO queue. Using a LIFO queue gave very unbalanced accesses to the TC blocks, there could
be as much as a 10x difference between the different TC blocks in a node.

Part of Mikael Ronstroms "Patches used in benchmark tree with Intel"