Fix Bug#13966453 CONCURRENT DML AND READS FROM I_S TABLES CRASH FETCH_DATA_INTO_CACHE_LOW
Traversing the list trx_sys->mysql_trx_list should be done like:

  for (trx = UT_LIST_GET_FIRST(trx_sys->mysql_trx_list);
       trx != NULL;
       trx = UT_LIST_GET_NEXT(mysql_trx_list, trx)) {

and not like:

  for (trx = UT_LIST_GET_FIRST(trx_sys->mysql_trx_list);
       trx != NULL;
       trx = UT_LIST_GET_NEXT(trx_list, trx)) {

This bug was introduced in:

  * revision-id: sunny.bains@oracle.com-20111120202539-tsreibci7v0r92e0
  * committer: Sunny Bains <Sunny.Bains@Oracle.Com>
  * branch nick: trunk
  * timestamp: Mon 2011-11-21 07:25:39 +1100
  * message:
  *   WL#6046 - Special handling of read-only transactions
  * ...

Before that changeset the code only traversed trx_sys->trx_list like:

  for (trx = UT_LIST_GET_FIRST(trx_sys->trx_list);
       trx != NULL;
       trx = UT_LIST_GET_NEXT(trx_list, trx)) {

and it did not fiddle with trx_sys->mysql_trx_list.

In WL#6046 the code was changed so that trx_sys->trx_list in the
initialization of the "for" loop became a parameter
(trx_sys->mysql_trx_list is passed as that parameter), but the update
statement of the "for" loop was not adjusted and was left like
"trx = UT_LIST_GET_NEXT(trx_list, trx)".