BUG#14134662: MUCH OVERHEAD FOR MD5() AND SHA1() ON SHORT STRINGS
BACKGROUND:
MySQL's hash functions MD5 and SHA relied on the somewhat slow 
sprintf function to convert the digests to hex representations.
This patch replaces the sprintf with a specific and inline hex
conversion function.

FIX:
This bug is fixed by using the same solution that was used
to fix Bug#11757443.