Bug#14031353 LOCK/ALGORITHM CLAUSES MISSING FROM              CREATE/DROP INDEX STATEMENTS
In WL#5534, ALTER TABLE statement was extended with a
LOCK clause (which controls concurrency) and an ALGORITHM
clause (which selects between in-place and copy ALTER).

The problem was that these two clauses, which are currently
most useful for index creation, were not also added to the
CREATE INDEX and DROP INDEX statments.

This patch corrects this oversight, by allowing the LOCK
and ALGORITHM clauses (comma separated) to be used for CREATE
and DROP INDEX as well.

The syntax is changed by adding
  [lock_algorithm_clause [lock_algorithm_clause]]
to the end of the current CREATE/DROP INDEX syntaxes.

lock_algorithm_clause:
    LOCK [=] {DEFAULT | NONE | SHARED | EXCLUSIVE}
  | ALGORITHM [=] {DEFAULT | INPLACE | COPY}

Test case added to alter_table.test.