WL#5980  - Allow remote tablespaces
This patch implements the following;
1)  Two new system tables, SYS_TABLESPACES & SYS_DATAFILES.
Code to create them, add and delete records when tablespaces are created
and dropped, and code to read them.
2)  Two new INFORMATION_SCHEMA tables;
INNODB_SYS_TABLESPACES & INNODB_SYS_DATAFILES
3)  New columns to INFORMATION_SCHEMA.INNODB_SYS_TABLES so that the
FLAGS field is interpreted like it is in INNODB_SYS_TABLESPACES.
4)  Code to use and display the DATA DIRECTORY clause of a CREATE TABLE
statement.
5)  Recovery mechanisms to find remote tablespaces.
6)  Changes to handle TRUNCATE TABLE, RENAME TABLE, DISCARD/IMPORT and
inplace alter.
7)  Code to allow a tablespace to be relocated.  When the engine starts up,
it reads the new location from an ISL file.
8)  Changes to several test cases and many new tests.

See worklog 5980 for a full description of the design.