Mats Kindahl	2012-05-29
Bug#14123372: SEMI-SYNC REPLICATION GETS WRONG NOTIFICATIONS AFTER A ROTATION
      
The set_trans_pos and get_trans_pos interface does not copy the file name and
just keep the pointer into the file name. The reason for this is to avoid the
overhead of managing memory for the copy. This causes a problem if a rotate is
executed between the set_trans_pos and the corresponding get_trans_pos.
      
This patch solves the problem by moving the replication hooks, which are the
users of this information, to before the rotation is executed. The commit hook
is placed inside ha_commit_low, which is called from all transation
coordinators, and the rollback hook is called from inside ha_rollback_low.
Since rotation cannot occur until all written transactions have been committed
in the engine, no rotation will occur before any transaction that is written to
the current binlog file have been committed or rolled back.