Fix Bug#14172780 CRASH OBSERVED ININNODB.INNODB-WL5522-DEBUG* TEST ON PB2 ON MULTIPLE PLATFORMS.
In dict_stats_analyze_index_level(), when we traverse the level we take care
to copy the rec into prev_rec when we are about to jump on the next page but
we forgot to do this in the case when the last rec on a page is delete-marked.
In this case we just "continue;", leaving prev_rec pointing somewhere inside
the old page which is going to be released by btr_pcur_move_to_next_user_rec()
when it jumps on the next page.

The fix is to also copy rec into prev_rec before "continue;"ing when a
delete-marked record is found and it is last on page and prev_rec is not
already copied (ie we may have this cond twice if a whole page is filled
with delete-marked records and prev_rec is (copied) from pageN where rec
jumps into pageN+1, which contains only delete-marked records and then rec
jumps into pageN+2 - no need to copy prev_rec again in this case).