BUG#14136654: ASSERTION FAILED ON BINLOG.WRITE_ONE_EMPTY_GROUP_TO_CACHE()
When a GTID was across multiple relay logs:
 RELAY LOG (X)
   GTID
   (EOF)
 RELAY LOG (X + 1)
   BEGIN
   INSERT
   COMMIT
it was wrongly interpreted as a empty group.
This happened when crash slave was used, on which configuration when 
PURGE logs is executed (on log rotate) a commit is done to ensure
consistency of relay info, that when presented with only a GTID on group
cache logged a empty group.

Fixed MYSQL_BIN_LOG::commit() method to correctly handle the case where
a GTID may start on relay log and end on different relay log file.