Fix for Bug#14167585 - LOTS OF TEST FAILURES ON TRUNK ON SOLARIS11/SPARC64
The test failures are caused by the YEAR(X) deprecation message not
being correctly printed when running on SPARC with 32 bit MySQL binaries.
Instead of printing "YEAR(X)" where X is the width of the YEAR column,
"YEAR(0)" is printed in the deprecation warning.

This is caused by that we store the width/length of the YEAR(X) in an
ulonglong variable in sql_yacc.yy. This will occupy 8 bytes on 32 bit
SPARC. But in the deprecation message (ER_INVALID_YEAR_COLUMN_LENGTH
in errmsg-utf8.txt) this is printed as "%lu" (ie. a 4 byte integer on
32 bit SPARC). Since SPARC is big-endian, this will cause the upper 4
bytes of the length variable to be printed. This will always be zero.

The fix for this problem is to replace ulonglong with ulong for
storing the length variable (in sql_yacc.yy). It is ok to do this
change since the length variable gets its value from the call to
strtoul() which returns an unsigned long.