Adjust index cardinality statistics calculation for WL#6255.
Do not skip statistics calculation if the table is being rebuilt online.
The outcome of the function dict_index_is_online_ddl() changes a bit in
WL#6255. Before WL#6255, it could only hold for secondary indexes.
After WL#6255, if dict_index_is_online_ddl() holds for a clustered index,
it means that the table is fully accessible, but it is merely being rebuilt
online and all changes to the table must be logged.

Index cardinality statistics calculation does not modify the table.
Therefore, it must continue "business as usual" for tables that happen
to be subjected to online rebuild.

This might have caused Valgrind warnings about reading uninitialized memory
in main.innodb_icp_all.