Fix for bug #14156617 "INPLACE ALTER: NOT SUPPORTED ALTER TABLE ... ADD PARTITION CAUSES CRASH".
Attempt to add a new partition to a partitioned NDB table led to crash
due to debug assertion in server built from mysql-trunk-cluster tree.
This also caused ndb_partition_range test failures in this tree.

The assertion failure occurred due to fact that prep_alter_part_table() 
has added new partition partition_info object for TABLE instance 
representing old version of table. Since NDB tables do not support 
"fast" changes of partitioning, after that we tried to carry out 
ALTER TABLE by copying data from old version of table to the new 
version. As in the process we tried to use TABLE object with modified
partition_info for reading data, which resulted in assertion failure 
due to discrepancy between real table structure and its description
in partition_info.

This patch solves the problem by ensuring that we don't modify 
partition_info object for TABLE instances representing old version
of table in cases when "fast" changes of partitioning is not going 
to be used. Instead we create a copy of partition_info, adjust it
and use for .FRM/.PAR files creation. Unfortunately we can't easily 
use this copy in all cases, since current "fast" partitioning API 
assumes that changes to partitioning are communicated to it through 
TABLE::part_info object.

There is no test case as bug is not repeatable without NDB (and
thus not repeatable in trunk) and mysql-trunk-cluster tree already 
contains ndb_partition_range.test which provides test coverage for 
this bug.