Patch updated according to suggestions from Magnus Blåudd and Fraqzer Clement.
This is a fix for bug#11748194 "TRANSACTION OBJECT CREATED AND UNRELEASED BY 
EXTRA CALL TO NEXTRESULT()".

This bug had the following effect: If an API application tried calling 
NdbScanOperation::nextResult() once more when the previous call had retuned
end-of-file (return code 1), the api would leak a transaction object.

This commit fixes that problem. Ndb::closeTransaction() will now put the
extra transaction object associated with the scan back in the idle list
for the right TC node.

Also, calling NdbScanOperation::nextResult() once too much will now set
a proper error code (4210), instead of the undefined code -1.

Finally, this commit adds a regression test that will trigger the bug if the
bugfix is not applied.