Bug#14215847 PURGE BINARY/MASTER LOGS: ASSERTION FAILED:              ! IS_SET(), FILE SQL_ERROR.CC
This assertion could be triggered by PURGE BINARY LOGS BEFORE
if the BEFORE clause contained an expression and an error
occured during the evaluation of this expression. PURGE would
then not abort as it should have, but continue execution
and eventually assert.

This patch fixes the problem by checking if the evaluation
of the expression fails and if so abort PURGE.