Bug#14208922 - RANGE OPTIMIZER SHOULD CONSIDER INDEX UNION                FOR DISJUNCTIONS INSIDE CONJUNCTIONS
When we have a condition of the form "WHERE (p1 or p2)"
where px is a predicate of the form indexed_col=const, 
the range optimizer calculates the cost of doing index merge 
union of the three indexes that covers px. However, if a 
simple range scan would have been an alternative to this 
index merge union, i.e. if the condition had been of the form 
"WHERE p0 AND (p1 or p2)", heuristics would be used to always
prefer a single-index range scan over the index covering p0.

The assumption that a single-index range is always cheaper 
than index merge union does not hold. If the selectivity of 
p0 is much lower than that of p1 and p2, index merge union 
should be chosen.

This changeset removes the heuristic and makes the range 
optimizer calculate both costs to be able to reach a cost based
decision of which range access method to pick.