Bug#14226481 FAILED ASSERTION 'STATE == TRX_STATE_NOT_STARTED'              IN TRX_STATE_EQ AT TRX0TRX.IC
The problem was that if DELETE with ORDER BY and subselect
caused a deadlock inside InnoDB, this deadlock was not
properly handled by the SQL layer. This meant that the
SQL layer would try to unlock the row after InnoDB had
rolled back the transaction. This caused an assertion
inside InnoDB.

This patch fixes the problem by checking for errors
inside find_all_keys() and not calling unlock_row()
if any errors have been reported by the storage engine.