BUG#11765698: MYISAM SELECTS LOCKS INSERTS INSIDE PROCEDURE ONLY BUG#11757947: QUERY CACHE INVALIDATED IN INSERT...SELECT IF NO ROWS               INSERTED
BACKGROUND:
When INSERTS to a large MyISAM table are coming from a stored
procedure and a SELECT query is also running concurrently 
INSERTS gets locked by the select which should not have happened.

When a normal INSERT is coming it does not get locked by the
SELECT statement which is correct.

ANALYSIS:
This bug can be fixed by reverting two (rather complete)
changes made by Sanja's fix for bug 14767 "INSERT in SF +
concurrent SELECTs may result in wrong data in query cache"
(revid: "sp1r-bell@book.local-20060315171552-49188").

Change#1 in sql_yacc.yy:
------------------------
The fix for bug 14767 upgraded the lock type from 
TL_WRITE_CONCURRENT_INSERT to TL_WRITE_DEFAULT for inserts
which were used in stored routines if we had a query cache
enabled. 
In other words, it disabled concurrent inserts from stored 
routines in this case. This was necessary to prevent a race 
condition, when a concurrent SELECT sneaked in between the 
point where query cache was invalidated by insert in stored 
routine and the moment when statement that has called this 
routine unlocked tables, making insert results visible to
other connections. 
In such a situation SELECT didn't see insert results, but 
still was stored in query cache and stayed there even after 
unlock, making SELECTs which followed to return stale results.

The above change for bug 14767 is no longer required, as it
was obsoleted by Davi's patch for bug 41098 "Query Cache
returns wrong result with concurrent insert" (revid:
"davi.arnaut@sun.com-20090219210935-9vilvcisyyieffxl").
The latter disables caching of SELECTs for tables which
have some concurrent inserts going on, thus making race
described above impossible. So there is no need for disabling
concurrent inserts from stored routines anymore.


Change#2 in sql_parse.cc:
------------------------
The fix for bug 14767 also tried to address a race, similar 
to the one described above, which occurred for INSERT ... SELECT,
in case when concurrent SELECT sneaked in between 
query_cache_invalidate3(), which happens in select_insert::send_eof(),
and unlocking of tables  which back then has happened in 
dispatch_command(). 
This was done by calling for INSERT ... SELECTs which used concurrent
inserts additional query_cache_invalidate() that was immediately 
followed by mysql_unlock_tables() (i.e. by making window for race
condition smaller).

Since fix for bug 41098 addresses this scenario as well (and in more
correct way) this part of patch also became redundant.

NOTE:
-------------------------------------------------------------------
Additionally, if we remove this change#2 this bug
BUG#11757947: QUERY CACHE INVALIDATED IN INSERT...SELECT IF NO ROWS 
              INSERTED
is also handled automatically.

BACKGROUND OF BUG#11757947:
When the query cache is enabled and an INSERT...SELECT query is 
executed that does not actually insert any rows, the query cache
entries for the table mentioned in the INSERT still get invalidated.

This is unnecessary and makes the query cache less effective.

ANALYSIS:
Change#2 made by the fix for bug 14767 is cause for this bug. 
Since this change#2 can be safely reverted as mentioned above,
this bug is handled automatically.
 
FIX OF BUG#11757947:
Change#2 has been removed since this is no longer required as the
patch for bug 41098 obsoletes the race mentioned above as well.
------------------------------------------------------------------

FIX FOR BUG#11765698:
From above discussion, we can safely revert complete patch of bug
14767. However, one of the change from this patch is just refactoring
a piece of code in sql_insert.cc file, which we can retain.

The effect of this fix is to make inserts always concurrent with 
selects.