Bug#13980947:rqg_opt_no_subquery_compare_51 fails with              STATUS_LENGTH_MISMATCH_SELECT on PB2.
PROBLEM:
While executing a query involving a join, less number of                                
tuples were fetched.
  
ANALYSIS:                                                                               
When there is a DYNAMIC RANGE scan involved in the query,                               
we do not set key_read to TRUE during execution. This is 
because, DYNAMIC_RANGE can be sometimes covering and 
sometimes not. (See Bug#14055700 for detailed explanation)
                                  
While moving the set_keyread call from                                                  
test_if_skip_sort_order to join_init_read_record as  
part of fix for the Bug#12713907, this was not taken care.                              
                                                                                        
FIX: 
Allow QUICK_RANGE_SELECT to decide if set_keyread                                       
should be called or not. So, call to set_keyread is made from                           
QUICK_RANGE_SELECT::reset() function if the index chosen is                             
covering.                                                                               
Having this call made from join_init_read_record() makes the code                       
complex with too many conditons to be checked and also its better                       
to allow each QUICK to decide if set_keyread needs to                                   
be called or not.