Bug#13735050: CRASH IN END_WRITE WHEN SUBQUERY               GROUPS BY OUTER FIELDS
PROBLEM:
There are two problems with respect to the query in case
of a debug build. One is the assert that happens in
resolve_ref_in_select_and_group and the other is a dump
in copy_funcs function.
The second part of the problem is fixed with
the patch for Bug#13966809. So we are resolving the
first dump happening in resolve_ref_in_select_and_group.

ANALYSIS:
While resolving a column reference in
resolve_ref_in_select_and_group, we see that the item being
referred to is not fixed and hence assert.  But the item
being referred in this case is an outer_ref item and outer_ref
items get fixed only in the call to fix_inner_refs which
happens later. Hence we need to check if the item is an
outer_ref and assert only if it is not.

FIX:
Do not assert if the item being referred to is an outer reference