Bug #13007062 : XML LIMITATION
The current XML parser implementation can only parse XML data
if the sum of the lengths of the nested tags increased by 1
doesn't exceed 128.
Otherwise "XML to large" warning is triggered and a NULL is 
returned.
This was because of the automatic buffer used in the parser state
to collect the nesting path.
Fixed by adding a dynamic overflow buffer when the automatic
buffer is depleted.
Added a realloc for the string library
Test case added.