Bug#14081638: Missing data on inner join + nested where with firstmatch
The FirstMatch semi-join strategy produced wrong results for a query
with the following plan:
 ot1 - it1 - nt1 - it2.

(The query that fails has three inner tables, but conceptually two
inner tables are sufficient to provoke the error.)
The plan has a semi-join nest consisting of tables it1 and it2, and when
creating the plan, the optimizer makes a "split jump" plan by putting
the non-dependent outer table nt1 between the two subquery tables.
However, execution for this type of plan was broken: Table it1 has a
jump back to table ot1, but this is based on existence of a row in it1,
even though it should be based on existence of a row in it2.

The problem is fixed by re-using the loosescan_match_tab field for
the FirstMatch semi-join strategy. For such "split-jump" strategy,
use it to check the found_match flag of the last inner table of the
semi-join. For FirstMatch without "split-jump", the last inner table is
the same as the table deciding whether to jump back to the outer table.

Performance overhead is extremely small: Checking of one extra flag
for the table deciding whether to perform a FirstMatch shortcut.

Some refactoring was also made: loosescan_match_tab was renamed to
match_tab because it no longer belongs exclusively to LooseScan.
Some inline access functions were created so that checking for
semi-join strategies during plan execution could be abstracted.

internal/mysql-test/suite/i_main/t/subquery.test
  Test case for bug#14081638

internal/mysql-test/suite/i_main/r/subquery.result
  Test case results for bug#14081638

sql/opt_explain.cc
  Adapted to new field and function names.

sql/sql_executor.cc
  In evaluate_join_record() moved FirstMatch check to after visiting
  following tables, so that join_tab->match_tab->found_match can be checked.
  Adapted to new field and function names.

sql/sql_select.cc
  In setup_semijoin_dups_elimination(), add setting of match_tab field
  for FirstMatch strategy.
  Adapted to new field and function names.

sql/sql_select.h
  Implemented inline functions do_firstmatch(), do_loosescan(),
  starts_weedout() and finishes_weedout() for class JOIN_TAB.
  Renamed field do_firstmatch to firstmatch_return.
  Renamed field loosescan_match_tab to match_tab.