Bug#13355430 ASSERT "! IS_SET() || M_CAN_OVERWRITE_STATUS"              DURING THD::CLEANUP => MY_DELETE()
This assert could occur when a client disconnected from the
server or during shutdown of the server. This was only
a problem on debug builds, no negative consequences on
release builds.

The assert is triggered if an error is reported to the
diagnostic area when it is non-empty (contains OK).
It was triggered in this case because we don't empty the
diagnostic area before we try to end a client connection.
So if the last statement before the disconnect ended with
OK, the assert would be triggered by any errors during
disconnect. In this specific case, the error was failure
to delete temporary table files.

This patch fixes the problem by resetting the diagnostics
area before cleaning up and ending a connection thread.