Bug#14234581 - SELECT ON INNODB TEMPORARY TABLE RETURNS ERROR 1412 - (ER_TABLE_DEF_CHANGED)
A temporary table is local to the session that created it. We rely on MySQL
for enforcing the visibility rules. Allow the transaction to open a temporary
table that has a transaction id greater than itself. The transaction is greater
because the table was created in a separate transaction after the current
transaction was already started. Therefore the new table is by definition
outside its consistent read view.

rb://1124 Approved by Jimmy Yang.