Bug #14304973 "IN-PLACE ALTER TABLE INCORRECTLY HANDLES PREFIX KEYS".
New in-place ALTER TABLE implementation tried to treat any ALTER on a
table with a prefix key as dropping and then re-creating such a key.
For engines that doesn't support in-place addition or dropping of
indexes this meant that any ALTER TABLE on such a table was carried
out using copy algorithm.
For engines that support in-place index operations this resulted in
additional unnecessary work when table with prefix key was altered.

The problem was that code in fill_alter_inplace_info() assumed that
it can determine if new and old version of a key use same field by
comparing pointers to Field objects from Create_field::field and
KEY_PART_INFO::field. This assumption was wrong for prefix keys
since for them KEY_PART_INFO::field points to a clone of Field
object with length adjusted to prefix length.

This fix changes the code to use comparison of field indexes
instead of pointers to Field objects.