BUG#14221840: KILLING HELP COMMAND: ASSERTION FAILED: ! IS_SET(),               FILE SQL_ERROR.CC, LINE 465
Analysis:
--------

MySqld crashes at times on a debug build, when the HELP command
is killed using KILL QUERY statement.

The crash is observed under a timing window where:

connection1: Run the HELP statement.

connection2: KILL the HELP query running on connection1. This sets
             the connection1 THD 'killed' status as 'KILL_QUERY'.

connection1: Since the THD is killed, a kill message is sent while
             attempting to read the help records. This sets the
             status of the diagnostic area to DA_ERROR.

At the end of execution of the HELP statement, the status in the
diagnostic area is set to DA_EOF only if no errors are reported
(DA_EMPTY). On a debug build, MySqld asserts if the status of the
diagnostic area is not DA_EMPTY. Since the status of diagnostic
area is set to DA_ERROR in the case mentioned above, MySqld crashes.
[On a release build, an appropriate error message ER_QUERY_INTERRUPTED:
Query execution interrupted is reported].

Fix:
---

Check the THD 'killed' status prior to setting the EOF status in the
diagnostic area. If the THD is killed, skip setting the EOF status
and perform the query cleanup.