BUG#13939940: SESSION-LEVEL LOW_PRIORITY_UPDATES DOES NOT               WORK FOR INSERT
BACKGROUND:
As Per documentation low_priority_updates variable/option lowers
priority of table modifying operations INSERT, UPDATE and DELETE
as compared to SELECT queries. However, it was not working for
INSERT statements (while it works fine for UPDATE and DELETE).

ANALYSIS:
At the parsing stage for tables to be modified by UPDATE or DELETE
statements TL_WRITE_DEFAULT lock type is set. Before tables are
locked, this type of lock is converted either to TL_WRITE or
TL_WRITE_LOW_PRIORITY, depending on session's low_priority_updates
value.

For INSERT statement TL_WRITE_CONCURRENT_INSERT type of lock is set
at the parsing stage and no similar conversion to TL_WRITE_LOW_PRIORITY
happens depending on low_priority_updates value.

FIX:
This bug is fixed by introducing a new type of lock similar
to TL_WRITE_DEFAULT -- TL_WRITE_CONCURRENT_DEFAULT, which is
used in INSERT if no explicit lock type was specified,
and is converted to TL_WRITE_CONCURRENT_INSERT or
TL_WRITE_LOW_PRIORITY depending on low_priority_updates value,
similarly to how it is done for TL_WRITE_DEFAULT.
NOTE:
Introducing a new lock type is required to override the effect
of low_priority_updates when CONCURRENT is explicitly specified
in the statement. For e.g LOAD DATA CONCURRENT, here 
overrides the effect of low_priority_updates. 
This is done to adhere to the principle that locks which are 
explicitly specified in the statement override locks which are 
implicitly prescribed by low_priority_updates. 
Quoting MySQL manual:
"If you specify HIGH_PRIORITY, it overrides the effect of the
--low-priority-updates option if the server was started with
that option."