Bug#14063233 ALTER INPLACE DOES NOT SUPPORT AUTO-PARTITIONED ENGINES
This patch updates the in-place ALTER implementation introduced
by WL#5534 to support auto-partitioning in engines such as NDB.

It adds a series of flags for ADD PARTITION, DROP PARTITION,
ALTER PARTITION etc. which will be communicated to the storage
engine for such ALTER TABLE operations.

It also expands the Alter_inplace_info class which describes the
changes to be done by ALTER TABLE to contain a partition_info object
describing pending partition changes. This makes it possible for
implementations of handler::check_if_supported_inplace_alter() to
take this object into consideration when deciding of ALTER TABLE
can be executed in-place.

Parts of the patch are written by Martin Sköld.

The patch does not add a test case as this problem is not repeatable
without NDB.