Patch for Bug#14317442: Regression failure when using xpath variable inside a stored procedure.
The user-visible problem was that inside a stored program, references to
stored program variables in XML-functions (like ExtractValue())
did not work after first execution of the stored program.

The technical problem was that parsing context and the sp-head object being executed,
were not properly set in sp_instr_stmt before calling mysql_execute_command().
So, XML-functions were unable to resolve stored program variables.

The fix is to set those attributes properly as it is done in sp_lex_instr::parse_expr()
for parsing separate expressions.