Fix Bug#14296397 INNODB.INNODB-WL5522-DEBUG TEST FAILS WITH TIMEOUT ON PB2 SPORADICALLY.
Synchronize the background stats gathering thread with
ALTER TABLE ... DISCARD TABLESPACE.

The problematic scenario is this:

STATS THREAD: (via btr_get_size())
x-latches the root page

DISCARD THREAD:
acquire the tablespace latch

STATS THREAD: (via btr_get_size())
try to acquire the tablespace latch (hangs, waiting on discard thread)

DISCARD THREAD:
wait on all pages from the tablespace to be unfixed in
buf_LRU_remove_all_pages(); this wait never ends because the stats thread
will never release the root page since it has hung trying to acquire the
tablespace latch.

The solution is to hook a "wait for bg stats thread to stop using this table"
in discard tablespace codepath. We know bg stats will not pick up again the
table after we have waited for it to stop using it because it only picks up
tables under data dictionary latch and the discard is holding that latch for
the duration where it acquires the tablespace latch and waits on all pages
to be unfixed.

Adding mtr testcase for this is not possible because DEBUG_SYNC() does not
work for background threads.

Approved by:	Sunny (via IM)