Bug#13813064 QUERY USING SPATIAL INDEX RETURNS              EMPTY RESULT Bug#14096685 WITHIN() YIELDS NO RESULTS WHEN              SPATIAL INDEX IS USED
The ST_Contains & Within() functions doesn't
show correct result when executed with spatial
index.

The reason behind this problem is that the spatial
index for Contains and Within class of functions in
MyISAM is implemented in opposite order.

Fixed by adjusting the flag that controls which
implementation to use and item's rev_functype()
for Contains & Within functions.