Bug#11766271 - ABORT IN BTR_PCUR_RESTORE_POSITION_FUNC
This bug is same as Bug #13510739. Fix for that bug is not
propagated to 5.6 & trunk.

We have an optimization to save CPU time: if index read is a
consistent read on a unique condition on the clustered index,
then we do not store the cursor position.

But this is not applicable for MYSQL HANDLER queries where the
user can move the cursor with PREV or NEXT even after a unique
search. So we disable the optimization i.e by storing the cursor
position even if the read is on a unique condition on the
clustered index.

Approved by Jimmy: rb://1168