BUG#14212892 - WHY DOES BUF_PAGE_GET_GEN CONFIRM THE PAGE                CHECKSUM FOR COMPRESSED PAGES
For compressed pages, page checksum verification is done
when reading the page from disk and also while decompressing
the compressed pages.

Since page checksum verification is already done when
the page is read from disk, it is not necessary to verify the
checksums again when decompressing the page. By removing the extra
page checksum verification, the CPU used will be saved.

No mtr testcase as this change is covered by existing testcases.

Approved by Inaam: rb://1166