Bug#14298586 - NANOSECOND TIMER NOT WORKING ON WINDOWS
Before this fix, the table performance_schema.setup_timers used always the
nanosecond timer for stages and statements, be default.

The problem with that is that this timer may not be available for the
platform used, and in this case, stages and statements timing is not
functioning properly out of the box.
For example, the NANOSECOND timer is not available on windows platforms.

With this fix, the idle, stage and statement timers are set to:
- use the preferred timers when available,
- use alternate timers when the preferred timers are not supported.

This is a major change for ease of use,
as it avoids the need to explicitly update table setup_timers.